introduce autotools machinery
A basic autotools machinery is introduced.
It is meant to work in parallel with the cmake machinery.
This first attempt builds a share library, against which
plfit and plgen are linked, and install a development
environment: systemwide headers, SO library, a pkg-config
metadata file.

To configure and build accordingly:
$ autoreconf -i
$ ./configure
$ make
$ make check
$ make install

To check the packaging:
$ make distcheck

To uninstall:
$ make uninstall

It should work fine in a _sane_ GNU environment.